requirements.txt
transformers>=4.38.0
trl>=0.7.10
peft>=0.10.0
bitsandbytes>=0.42.0
safetensors>=0.4.0
ragas>=0.1.7
gradio>=4.16.0
openai>=1.0.0
tqdm
pandas
scikit-learn
jsonlines
loguru
python-dotenv

# For CUDA support (if available)
torch>=2.1.0

# For minimal GPU, avoid large dependencies 